+++ include/NekiraReflect/DynamicReflect/DynamicReflect.hpp
@@ -25,7 +25,7 @@
 
 #pragma once
 
-
+#include <Macros/CoreMacros.hpp>
 #include <Utility/Utilities.hpp>
 
 // ============================================= 动态反射注册工具 ============================================= //
