+++ include/NekiraReflect/Generation/Generator/ReflectGenerator.hpp
@@ -25,8 +25,7 @@
 
 #pragma once
 
-#include <Tools/CodeGenerateHelper.hpp>
-#include <fstream>
+#include <Tools/CodeConsume.hpp>
 
 
 
@@ -36,32 +35,6 @@ namespace NekiraReflect
 class ReflectGenerator final
 {
 public:
-    // 生成枚举、类、结构体的反射代码
-    // OutputFile: 输出文件名(不包含后缀)
-    static void GenerateCode(const std::string& OutputFile, const std::vector<EnumMetaInfo>& Enums,
-                             const std::vector<ClassMetaInfo>& Classes)
-    {
-        const std::string HeaderName = OutputFile + ".gen.hpp";
-
-        std::ofstream HeaderStream(HeaderName);
-
-        HeaderStream << "#pragma once" << '\n' << '\n';
-        HeaderStream << "#include <NekiraReflect/DynamicReflect/Accessor/ReflectAccessor.hpp>" << '\n';
-        HeaderStream << '\n';
-
-        // 生成枚举的反射代码
-        if (!Enums.empty())
-        {
-            CodeGenerateHelper::GenerateEnumCode(HeaderStream, Enums);
-        }
-
-        // 生成类的反射注册代码
-        if (!Classes.empty())
-        {
-            CodeGenerateHelper::GenerateClassCode(HeaderStream, Classes);
-        }
-    }
-
 private:
 };
