+++ include/NekiraReflect/Generation/Tools/CodeGenerateHelper.hpp
@@ -34,6 +34,7 @@
 
 namespace NekiraReflect
 {
+
 class CodeGenerateHelper final
 {
 public:
